fix(models): preserve CLI package file-count baseline
